mtd: add leading underscore to all mtd functions
[backport from linux commit 3c3c10bba1e4ccb75b41442e45c1a072f6cded19]

This patch renames all MTD functions by adding a "_" prefix:

mtd->erase -> mtd->_erase
mtd->read_oob -> mtd->_read_oob
...

The reason is that we are re-working the MTD API and from now on it is
an error to use MTD function pointers directly - we have a corresponding
API call for every pointer. By adding a leading "_" we achieve the following:

1. Make sure we convert every direct pointer users
2. A leading "_" suggests that this interface is internal and it becomes
   less likely that people will use them directly
3. Make sure all the out-of-tree modules stop compiling and the owners
   spot the big API change and amend them.

This commit is part of the synchronization of the barebox mtd layer with
Linux v3.4.

Signed-off-by: Dirk Hoerner <dirk.hoerner@dspg.com>